+++ pulse-binding/src/mainloop/standard.rs
@@ -278,6 +278,8 @@ impl super::api::Mainloop for Mainloop {
     }
 }
 
+impl super::signal::MainloopSignals for Mainloop {}
+
 impl super::api::MainloopInner<MainloopInternal> {
     fn drop_actual(&mut self) {
         unsafe { capi::pa_mainloop_free(self.ptr) };